# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amber — Rare Disease Atlas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/path": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Find Path",
        "description": "Ordered path steps, or no_supported_route with a coverage report.",
        "operationId": "findPath",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start node ID.",
              "title": "From"
            },
            "description": "Start node ID."
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End node ID.",
              "title": "To"
            },
            "description": "End node ID."
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PathFamily",
              "default": "all"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 3,
              "title": "K"
            }
          },
          {
            "name": "include_vus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Allow VUS variants on paths.",
              "default": false,
              "title": "Include Vus"
            },
            "description": "Allow VUS variants on paths."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: lists invalid field names only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfidenceLevel": {
        "type": "string",
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "ConfidenceLevel"
      },
...
      "CoverageReport": {
        "properties": {
          "sources_queried": {
            "items": {
              "$ref": "#/components/schemas/SourceCount"
            },
            "type": "array",
            "title": "Sources Queried"
          },
          "closest_partial_path": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Path"
              },
              {
                "type": "null"
              }
            ],
            "description": "Best path found, even though it is below the threshold."
          },
          "missing_link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MissingLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "suggested_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Question",
            "description": "Suggested next question."
          }
        },
        "type": "object",
        "title": "CoverageReport"
      },
...
      "Edge": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Edge ID, 'e_' + 12 hex chars."
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "description": "Source node ID."
          },
          "target_id": {
            "type": "string",
            "title": "Target Id",
            "description": "Target node ID."
          },
          "relation": {
            "$ref": "#/components/schemas/Relation",
            "description": "Relation type."
          },
          "family": {
            "$ref": "#/components/schemas/EdgeFamily",
            "description": "Edge family: dna, symptoms, research, community."
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Stage 4 confidence, 0..1."
          },
          "confidence_level": {
            "$ref": "#/components/schemas/ConfidenceLevel",
            "description": "High >= 0.8, Medium >= 0.5, else Low."
          },
          "origin": {
            "$ref": "#/components/schemas/Origin",
            "description": "observed (data) vs inferred (hypothesis) and others."
          },
          "status": {
            "$ref": "#/components/schemas/EdgeStatus",
            "description": "active, pending_review, or under_review (open user flags)."
          },
          "features": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features",
            "description": "Features behind inferred edges (shared HPO terms, scores ...)."
          },
          "data_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version",
            "description": "Graph data version."
          },
          "evidence_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Evidence Count",
            "description": "Number of evidence rows.",
            "default": 0
          },
          "contradiction_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Contradiction Count",
            "description": "Evidence rows that contradict.",
            "default": 0
          },
          "flagged": {
            "type": "boolean",
            "title": "Flagged",
            "description": "True when the edge has open user flags.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "target_id",
          "relation",
          "family",
          "confidence",
          "confidence_level",
          "origin",
          "status"
        ],
        "title": "Edge"
      },
...
      "EdgeFamily": {
        "type": "string",
        "enum": [
          "dna",
          "symptoms",
          "research",
          "community"
        ],
        "title": "EdgeFamily"
      },
...
      "EdgeStatus": {
        "type": "string",
        "enum": [
          "active",
          "pending_review",
          "under_review"
        ],
        "title": "EdgeStatus"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "bad_request",
          "sign_in_required",
          "consent_required",
          "forbidden",
          "not_found",
          "conflict",
          "payload_too_large",
          "unsupported_media_type",
          "validation_error",
          "rate_limited",
          "not_implemented",
          "upstream_error",
          "reauth_required",
          "internal_error"
        ],
        "title": "ErrorCode"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message. Never echoes user content."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The single error envelope used by every endpoint."
      },
...
      "MissingLink": {
        "properties": {
          "from_id": {
            "type": "string",
            "title": "From Id"
          },
          "to_id": {
            "type": "string",
            "title": "To Id"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "What evidence is missing, in plain language."
          }
        },
        "type": "object",
        "required": [
          "from_id",
          "to_id",
          "description"
        ],
        "title": "MissingLink"
      },
...
      "Node": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable node ID, e.g. MONDO:0100135, HGNC:11444, HP:0001250."
          },
          "type": {
            "$ref": "#/components/schemas/NodeType",
            "description": "Node type."
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Display label."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Short description from the source."
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "Link to the authoritative source page."
          },
          "attrs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attrs",
            "description": "Type-specific attributes (e.g. variant classification, registry kind)."
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id",
            "description": "Mechanism/phenotype cluster, if any."
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X",
            "description": "Precomputed layout x position."
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y",
            "description": "Precomputed layout y position."
          },
          "centrality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Centrality",
            "description": "Precomputed centrality score (0..1)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "label"
        ],
        "title": "Node"
      },
...
      "NodeType": {
        "type": "string",
        "enum": [
          "disease",
          "gene",
          "variant",
          "mechanism",
          "pathway",
          "phenotype",
          "paper",
          "claim",
          "researcher",
          "doctor",
          "institution",
          "network",
          "grant",
          "trial",
          "patient_org",
          "registry",
          "cluster"
        ],
        "title": "NodeType"
      },
      "Origin": {
        "type": "string",
        "enum": [
          "observed",
          "inferred",
          "patient_reported",
          "user_contributed"
        ],
        "title": "Origin"
      },
      "Path": {
        "properties": {
          "path_id": {
            "type": "string",
            "title": "Path Id",
            "description": "path_id(edge_ids): 'p_' + 16 hex chars."
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PathStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "edge_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Edge Ids",
            "description": "Edge IDs in traversal order."
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost",
            "description": "Sum of -log(confidence) over the edges."
          },
          "min_confidence": {
            "type": "number",
            "title": "Min Confidence",
            "description": "Lowest edge confidence on the path."
          },
          "min_confidence_level": {
            "$ref": "#/components/schemas/ConfidenceLevel"
          },
          "all_observed": {
            "type": "boolean",
            "title": "All Observed",
            "description": "Every edge has origin 'observed'."
          },
          "all_active": {
            "type": "boolean",
            "title": "All Active",
            "description": "Every edge has status 'active'."
          },
          "supported": {
            "type": "boolean",
            "title": "Supported",
            "description": "Every edge passes the confidence threshold."
          }
        },
        "type": "object",
        "required": [
          "path_id",
          "steps",
          "edge_ids",
          "total_cost",
          "min_confidence",
          "min_confidence_level",
          "all_observed",
          "all_active",
          "supported"
        ],
        "title": "Path"
      },
      "PathFamily": {
        "type": "string",
        "enum": [
          "dna",
          "symptoms",
          "research",
          "all"
        ],
        "title": "PathFamily"
      },
      "PathResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PathStatus"
          },
          "from_id": {
            "type": "string",
            "title": "From Id"
          },
          "to_id": {
            "type": "string",
            "title": "To Id"
          },
          "family": {
            "$ref": "#/components/schemas/PathFamily"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold",
            "description": "Confidence threshold used.",
            "default": 0.6
          },
          "paths": {
            "items": {
              "$ref": "#/components/schemas/Path"
            },
            "type": "array",
            "title": "Paths",
            "description": "Top-k supported paths."
          },
          "coverage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoverageReport"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present when status is no_supported_route."
          }
        },
        "type": "object",
        "required": [
          "status",
          "from_id",
          "to_id",
          "family"
        ],
        "title": "PathResponse"
      },
      "PathStatus": {
        "type": "string",
        "enum": [
          "ok",
          "no_supported_route"
        ],
        "title": "PathStatus"
      },
      "PathStep": {
        "properties": {
          "edge": {
            "$ref": "#/components/schemas/Edge"
          },
          "from_node": {
            "$ref": "#/components/schemas/Node",
            "description": "Node the step starts at (in traversal order)."
          },
          "to_node": {
            "$ref": "#/components/schemas/Node",
            "description": "Node the step ends at (in traversal order)."
          },
          "reversed": {
            "type": "boolean",
            "title": "Reversed",
            "description": "True when traversed against the stored edge direction.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "edge",
          "from_node",
          "to_node"
        ],
        "title": "PathStep"
      },
...
      "Relation": {
        "type": "string",
        "enum": [
          "caused_by_variant_in",
          "acts_via",
          "participates_in",
          "has_phenotype",
          "same_gene_same_mechanism",
          "same_gene_different_mechanism",
          "shared_pathway",
          "similar_symptoms",
          "shared_researcher",
          "asserts",
          "authored",
          "pi_of",
          "funds_research_on",
          "serves",
          "runs",
          "studies",
          "investigator_of",
          "affiliated_with",
          "variant_of",
          "observed_in",
          "member_of",
          "about"
        ],
        "title": "Relation"
      },
...
      "SourceCount": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source name, e.g. ClinVar, HPO, PubMed."
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of results the source contributed."
          }
        },
        "type": "object",
        "required": [
          "source",
          "count"
        ],
        "title": "SourceCount"
      },
...
    }
  }
}